+++ src/sr_olthad/__init__.py
@@ -1,8 +1,8 @@
 from sr_olthad.lm_step import (
-    GetDomainSpecificInsert,
     PostLmStepApprover,
     PreLmStepHandler,
 )
 from sr_olthad.sr_olthad import SrOlthad
+from sr_olthad.utils import GetDomainSpecificInsert
 
 # TODO: Import other stuff like important types that users should know about?
