+++ src/sr_olthad/sr_olthad.py
@@ -4,9 +4,9 @@ from collections.abc import Callable
 import sr_olthad.config as cfg
 from common.agents import LmRetryHandler
 from common.schema import LmStreamsHandler
+from sr_olthad import GetDomainSpecificInsert
 from sr_olthad.agents import AttemptSummarizer, Backtracker, Forgetter, Planner
 from sr_olthad.lm_step import (
-    GetDomainSpecificInsert,
     LmStepTemplate,
     PostLmStepApprover,
     PreLmStepHandler,
